move tests into main file
